requirements.txt
exifread
iptcinfo3
ninepatch
numpy
paho-mqtt
pi3d
pyyaml
urllib3